+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ Release date: ``2020-xx-xx``
 
 - `NXPY-159 <https://jira.nuxeo.com/browse/NXPY-159>`__: Allow to pass additional arguments to ``Batch.complete()``
 - `NXPY-145 <https://jira.nuxeo.com/browse/NXPY-145>`__: Detect and log appropriate debug info when the transfer if chunked
+- `NXPY-163 <https://jira.nuxeo.com/browse/NXPY-163>`__: Add the capability to refresh tokens in third-party batch handlers
 - `NXPY-164 <https://jira.nuxeo.com/browse/NXPY-164>`__: Clean-up code smells found by Sourcery
 
 Technical changes
@@ -16,6 +17,7 @@ Technical changes
 - ``Batch.complete()`` now handles additional parameters
 - Added nuxeo/constants.py::\ ``LOG_LIMIT_SIZE``
 - nuxeo/uploads.py::\ ``API.complete()`` now handles additional parameters
+- Added nuxeo/uploads.py::\ ``API.refresh_token()``
 
 2.4.4
 -----